requirements.txt
langchain
python-dotenv
streamlit
unstructured
tiktoken
faiss-cpu
libmagic
python-magic
python-magic-bin
OpenAI
langchain_groq
langchain_community
langchain_huggingface